+++ examples/copyContentToProjectFolder/applications/example/app/examples/simple/jsonpGet.jsonp/get.js
@@ -1,0 +1,7 @@
+﻿module.exports = function(client, callback) {
+  callback({
+    query:  client.query,
+    path:   client.path,
+    fields: client.fields
+  });
+}