+++ HISTORY.md
@@ -3,6 +3,7 @@
 
   * Improve AJAX handlers, close #169
   * Middleware functions: application.get(path, handler, meta), .post, .put, .delete, close #168, related issue #165 and #165
+  * Some code refactoring
 
 0.1.157 / 2014-11-27
 ==================